feat: use fts5 for search
The data for locations is often not great: it mixes different sub-areas
for a place with numbers like. For example for "Guadalupe Victoria", I
often see:

  - "Guadalupe Victoria (sector 1, 2ª sección Oeste)",
  - "Guadalupe Victoria (2ª sección sector 1)",
  - "Guadalupe Victoria (sector 5 parte baja)", and of course
  - just "Guadalupe Victoria".

I have about 118 different location names in the DB for now. This
number will grow, and once it is stabilized, maybe I'll take a look at
improving the existing schema with a proper "locations" table, etc. (Yay
data migration!).

Right now, I can only make guesses as to how names will be formatted, so
the data cleaning can wait.